requirements.txt
asgiref==3.2.7
asn1crypto==1.3.0
attrs==19.3.0
Babel==2.8.0
certifi==2020.4.5.1
# cffi==1.14.0
chardet==3.0.4
cryptography==2.9
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.5
django-cities==0.6
django-cors-headers==3.2.1
django-model-utils==4.0.0
django-phonenumber-field==4.0.0
django-rest-framework==0.1.0
django-versatileimagefield==2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
facebook-sdk==3.1.0
future==0.18.2
gunicorn==20.0.4
idna==2.6
importlib-metadata==1.6.0
JSON-minify==0.3.0
jsonschema==3.2.0
# lxml==4.5.0
oauthlib==3.1.0
olefile==0.46
opencivicdata==3.2.0
phonenumbers==8.12.1
Pillow==7.1.1
plivo==4.6.1
psycopg2-binary==2.8.5
pupa==0.10.1
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyopenstates==1.2.0
pyrsistent==0.16.0
python-magic==0.4.15
python-twitter==3.5
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
scrapelib==1.2.0
selenium==3.141.0
six==1.14.0
spatula==0.0.2
SQLAlchemy==1.3.16
sqlparse==0.3.1
swapper==1.1.2.post1
tqdm==4.45.0
urllib3==1.25.9
xvfbwrapper==0.2.9
zipp==3.1.0

